requirements.txt
python-can
cantools
tqdm